bugfix: fix comparison of fields in schema and struct
This function assumes the order of fields in the schema and the corresponding
struct to the be same. Even though fields are present in one or the other
this function will mark it as invalidate.

Therefore, we create a map and query the schemas and match them and then
check for validity